Adjust asynchronous diagnostic job and helpers
The helper classes AsyncDiagHelper and AsyncDiagJobHelper
were once introduced in order to make "long running"
diagnostic jobs (i.e. jobs that don't send the result
synchronously from process) multi-tester capable: each
such job can contain an instance of AsyncDiagJobHelper that
keeps state whether a job is already in progress and
enqueues new requests and delays processing of any new
request until the running job is executed completely.

This functionality was intended to operate only on diagnostic
task and when async OS abstraction was introduced the context
of the diagnostic task was added as a parameter. In the
scope of AsyncDiagJob this parameter was misinterpreted and
from then on AsyncDiagJob instances were used to execute
diag jobs asynchronously in arbitrary OS tasks. While the
AsyncDiagJobHelper class (member of AsyncDiagJob) in this
case didn't get the context of the diagnostic task but also
this arbitrary task this could lead to subtle consistency
problems when executing diagnostic jobs.

In order to avoid such problems move the context of the
diagnostic task AsyncDiagHelper (there's only one
instance per application) and remove it as parameter to
constructor call of AsyncDiagJobHelper. Instead it is now
requested from AsyncJobHelper.

Change-Id: I4e36e1f300011bea47838ba957ce820d881e1552